[ui] Refactor RunTable into more reusable components
