+++ js_modules/dagster-ui/packages/ui-core/src/runs/RunRow.tsx
@@ -1,38 +1,20 @@
-import {
-  Box,
-  Button,
-  ButtonLink,
-  Caption,
-  Checkbox,
-  Colors,
-  Dialog,
-  DialogBody,
-  DialogFooter,
-  Mono,
-  Tag,
-} from '@dagster-io/ui-components';
+import {Box, ButtonLink, Caption, Checkbox, Colors, Mono, Tag} from '@dagster-io/ui-components';
 import * as React from 'react';
 import {Link} from 'react-router-dom';
 import styled from 'styled-components';
 
-import {AssetCheckTagCollection, AssetKeyTagCollection} from './AssetTagCollections';
 import {CreatedByTagCell} from './CreatedByTag';
 import {QueuedRunCriteriaDialog} from './QueuedRunCriteriaDialog';
 import {RunActionsMenu} from './RunActionsMenu';
+import {RunRowTags} from './RunRowTags';
 import {RunStatusTagWithStats} from './RunStatusTag';
-import {DagsterTag, TagType} from './RunTag';
-import {RunTags} from './RunTags';
-import {RunStateSummary, RunTime, assetKeysForRun, titleForRun} from './RunUtils';
-import {RunFilterToken, runsPathWithFilters} from './RunsFilterInput';
+import {DagsterTag} from './RunTag';
+import {RunTargetLink} from './RunTargetLink';
+import {RunStateSummary, RunTime, titleForRun} from './RunUtils';
+import {RunFilterToken} from './RunsFilterInput';
 import {RunTableRunFragment} from './types/RunTable.types';
-import {useTagPinning} from './useTagPinning';
-import {ShortcutHandler} from '../app/ShortcutHandler';
-import {isHiddenAssetGroupJob} from '../asset-graph/Utils';
-import {PipelineTag, RunStatus} from '../graphql/types';
-import {PipelineReference} from '../pipelines/PipelineReference';
-import {buildRepoAddress} from '../workspace/buildRepoAddress';
-import {RepoAddress} from '../workspace/types';
-import {useRepositoryForRunWithoutSnapshot} from '../workspace/useRepositoryForRun';
+import {useResolveRunTarget} from './useResolveRunTarget';
+import {RunStatus} from '../graphql/types';
 
 export const RunRow = ({
   run,
@@ -57,26 +39,7 @@ export const RunRow = ({
   isHighlighted?: boolean;
   hideCreatedBy?: boolean;
 }) => {
-  const {pipelineName} = run;
-  const repo = useRepositoryForRunWithoutSnapshot(run);
-  const {isTagPinned, onToggleTagPin} = useTagPinning();
-
-  const isJob = React.useMemo(() => {
-    if (repo) {
-      const pipelinesAndJobs = repo.match.repository.pipelines;
-      const match = pipelinesAndJobs.find((pipelineOrJob) => pipelineOrJob.name === pipelineName);
-      return !!match?.isJob;
-    }
-    return false;
-  }, [repo, pipelineName]);
-
-  const repoAddressGuess = React.useMemo(() => {
-    if (repo) {
-      const {match} = repo;
-      return buildRepoAddress(match.repository.name, match.repositoryLocation.name);
-    }
-    return null;
-  }, [repo]);
+  const {isJob, repoAddressGuess} = useResolveRunTarget(run);
 
   const onChange = (e: React.FormEvent<HTMLInputElement>) => {
     if (e.target instanceof HTMLInputElement) {
@@ -87,63 +50,16 @@ export const RunRow = ({
     }
   };
 
-  const allTagsWithPinned = React.useMemo(() => {
-    const allTags: Omit<PipelineTag, '__typename'>[] = [...run.tags];
-    if ((isJob && run.mode !== 'default') || !isJob) {
-      allTags.push({key: 'mode', value: run.mode});
-    }
-    return allTags.map((tag) => {
-      return {...tag, pinned: isTagPinned(tag)};
-    });
-  }, [run, isJob, isTagPinned]);
-
   const isReexecution = run.tags.some((tag) => tag.key === DagsterTag.ParentRunId);
 
-  const [showRunTags, setShowRunTags] = React.useState(false);
   const [showQueueCriteria, setShowQueueCriteria] = React.useState(false);
   const [isHovered, setIsHovered] = React.useState(false);
-
-  const tagsToShow = React.useMemo(() => {
-    const targetBackfill = allTagsWithPinned.find((tag) => tag.key === DagsterTag.Backfill);
-    const tagKeys: Set<string> = new Set([]);
-    const tags: TagType[] = [];
-
-    if (targetBackfill && targetBackfill.pinned) {
-      const link = run.assetSelection?.length
-        ? `/overview/backfills/${targetBackfill.value}`
-        : runsPathWithFilters([
-            {
-              token: 'tag',
-              value: `${DagsterTag.Backfill}=${targetBackfill.value}`,
-            },
-          ]);
-      tags.push({
-        ...targetBackfill,
-        link,
-      });
-      tagKeys.add(DagsterTag.Backfill as string);
-    }
-    allTagsWithPinned.forEach((tag) => {
-      if (tagKeys.has(tag.key)) {
-        // We already added this tag
-        return;
-      }
-      if (tag.pinned) {
-        tags.push(tag);
-      }
-    });
-    return tags;
-  }, [allTagsWithPinned, run.assetSelection?.length]);
 
   return (
     <Row
       highlighted={!!isHighlighted}
-      onMouseEnter={() => {
-        setIsHovered(true);
-      }}
-      onMouseLeave={() => {
-        setIsHovered(false);
-      }}
+      onMouseEnter={() => setIsHovered(true)}
+      onMouseLeave={() => setIsHovered(false)}
     >
       {hasCheckboxColumn ? (
         <td>
@@ -174,24 +90,8 @@ export const RunRow = ({
             flex={{direction: 'row', alignItems: 'center', wrap: 'wrap'}}
             style={{gap: '4px 8px'}}
           >
-            <RunTagsWrapper>
-              {tagsToShow.length ? (
-                <RunTags tags={tagsToShow} onAddTag={onAddTag} onToggleTagPin={onToggleTagPin} />
-              ) : null}
-            </RunTagsWrapper>
-            {allTagsWithPinned.length > tagsToShow.length ? (
-              <Caption>
-                <ButtonLink
-                  onClick={() => {
-                    setShowRunTags(true);
-                  }}
-                  color={Colors.textLight()}
-                  style={{margin: '-4px', padding: '4px'}}
-                >
-                  View all tags ({allTagsWithPinned.length})
-                </ButtonLink>
-              </Caption>
-            ) : null}
+            <RunRowTags run={run} isJob={isJob} isHovered={isHovered} onAddTag={onAddTag} />
+
             {run.status === RunStatus.QUEUED ? (
               <Caption>
                 <ButtonLink
@@ -206,18 +106,6 @@ export const RunRow = ({
             ) : null}
           </Box>
         </Box>
-        {isHovered && allTagsWithPinned.length ? (
-          <ShortcutHandler
-            key="runtabletags"
-            onShortcut={() => {
-              setShowRunTags((showRunTags) => !showRunTags);
-            }}
-            shortcutLabel="t"
-            shortcutFilter={(e) => e.code === 'KeyT'}
-          >
-            {null}
-          </ShortcutHandler>
-        ) : null}
       </td>
       {hideCreatedBy ? null : (
         <td>
@@ -242,29 +130,6 @@ export const RunRow = ({
           additionalActionsForRun={additionalActionsForRun}
         />
       </td>
-      <Dialog
-        isOpen={showRunTags}
-        title="Tags"
-        canOutsideClickClose
-        canEscapeKeyClose
-        onClose={() => {
-          setShowRunTags(false);
-        }}
-      >
-        <DialogBody>
-          <RunTags tags={allTagsWithPinned} onAddTag={onAddTag} onToggleTagPin={onToggleTagPin} />
-        </DialogBody>
-        <DialogFooter topBorder>
-          <Button
-            intent="primary"
-            onClick={() => {
-              setShowRunTags(false);
-            }}
-          >
-            Close
-          </Button>
-        </DialogFooter>
-      </Dialog>
       <QueuedRunCriteriaDialog
         run={run}
         isOpen={showQueueCriteria}
@@ -278,34 +143,3 @@ const Row = styled.tr<{highlighted: boolean}>`
   ${({highlighted}) =>
     highlighted ? `box-shadow: inset 3px 3px #bfccd6, inset -3px -3px #bfccd6;` : null}
 `;
-
-const RunTagsWrapper = styled.div`
-  display: contents;
-  > * {
-    display: contents;
-  }
-`;
-
-const RunTargetLink = ({
-  run,
-  isJob,
-  repoAddress,
-}: {
-  isJob: boolean;
-  run: RunTableRunFragment;
-  repoAddress: RepoAddress | null;
-}) => {
-  return isHiddenAssetGroupJob(run.pipelineName) ? (
-    <Box flex={{gap: 16, alignItems: 'end', wrap: 'wrap'}}>
-      <AssetKeyTagCollection assetKeys={assetKeysForRun(run)} />
-      <AssetCheckTagCollection assetChecks={run.assetCheckSelection} />
-    </Box>
-  ) : (
-    <PipelineReference
-      isJob={isJob}
-      showIcon
-      pipelineName={run.pipelineName}
-      pipelineHrefContext={repoAddress || 'repo-unknown'}
-    />
-  );
-};
